Change to handle models inheriting from user.
The old code blows up because source_field_name remains None.  To test this:

Create a model that inherits from User:
class Employee(User):
    pass

And an admin that also inherits UserAdmin:
class EmployeeAdmin(UserAdmin):
    pass
site.register(Employee, EmployeeAdmin)

Add it to your settings and then go add an employee through it.  Adding works fine.  The change view is displayed after adding.  Fill in the email or something and then save and it will blow up.  The reason is because instance is the Employee model which inherits.  It is an instance of User but it is not the same type.